moved encoding tests to new file
